+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+- Add convolutional encoder `modules.CNNEncoder` and channel layer normalization `modules.ChannelNorm`
+- Add tests for `models` module
+- Further refine linting and add module docstrings
+
 # 0.1.0
 - Boilerplate, documentation with `Sphinx`, linting with `Ruff`, unit testing
 - Initial commit